Improve form accessiblity and responsiveness
Associate labels and inputs appropriately for accessiblity purposes.
Include basic validation for email and subject fields.

Style textarea so that it can only be resized vertically, and there is a
specified maximum and minimum height that the user can adjust it to.

Further iterations could improve the default form validation, and
provide a submission handler.